+++ src/aux-server/aux-backend/prisma/PrismaPolicyStore.ts
@@ -3,6 +3,7 @@ import {
     DEFAULT_ANY_RESOURCE_POLICY_DOCUMENT,
     DEFAULT_PUBLIC_READ_POLICY_DOCUMENT,
     GetUserPolicyResult,
+    ListUserPoliciesStoreResult,
     ListedRoleAssignments,
     ListedUserPolicy,
     PUBLIC_READ_MARKER,
@@ -55,7 +56,7 @@ export class PrismaPolicyStore implements PolicyStore {
     async listUserPolicies(
         recordName: string,
         startingMarker: string
-    ): Promise<ListedUserPolicy[]> {
+    ): Promise<ListUserPoliciesStoreResult> {
         let query: Prisma.PolicyWhereInput = {
             recordName: recordName,
         };
@@ -73,12 +74,19 @@ export class PrismaPolicyStore implements PolicyStore {
             take: 10,
         });
 
-        return policies.map((p) => {
+        const results = policies.map((p) => {
             return {
                 marker: p.marker,
                 document: p.document as unknown as PolicyDocument,
-            } as ListedUserPolicy;
+                markers: p.markers,
+            };
         });
+
+        return {
+            success: true,
+            policies: results,
+            totalCount: results.length,
+        };
     }
 
     async listRolesForUser(
@@ -173,32 +181,120 @@ export class PrismaPolicyStore implements PolicyStore {
             },
         });
 
+        const results = assignments.map((a) => {
+            if (a.type === 'inst') {
+                return {
+                    type: 'inst',
+                    inst: a.subjectId,
+                    role: {
+                        role: a.roleId,
+                        expireTimeMs: getExpireTime(
+                            convertToMillis(a.expireTime)
+                        ),
+                    },
+                } as const;
+            } else {
+                return {
+                    type: 'user',
+                    userId: a.userId,
+                    role: {
+                        role: a.roleId,
+                        expireTimeMs: getExpireTime(
+                            convertToMillis(a.expireTime)
+                        ),
+                    },
+                } as const;
+            }
+        });
+
         return {
-            assignments: assignments.map((a) => {
-                if (a.type === 'inst') {
-                    return {
-                        type: 'inst',
-                        inst: a.subjectId,
-                        role: {
-                            role: a.roleId,
-                            expireTimeMs: getExpireTime(
-                                convertToMillis(a.expireTime)
-                            ),
-                        },
-                    };
-                } else {
-                    return {
-                        type: 'user',
-                        userId: a.userId,
-                        role: {
-                            role: a.roleId,
-                            expireTimeMs: getExpireTime(
-                                convertToMillis(a.expireTime)
-                            ),
-                        },
-                    };
-                }
-            }),
+            assignments: results,
+            totalCount: results.length,
+        };
+    }
+
+    async listAssignments(
+        recordName: string,
+        startingRole: string
+    ): Promise<ListedRoleAssignments> {
+        const now = new Date();
+        const simpleQuery: Prisma.RoleAssignmentWhereInput = {
+            recordName: recordName,
+            OR: [
+                {
+                    expireTime: {
+                        gt: now,
+                    },
+                },
+                {
+                    expireTime: {
+                        equals: null,
+                    },
+                },
+            ],
+        };
+        let query: Prisma.RoleAssignmentWhereInput = {
+            ...simpleQuery,
+        };
+
+        if (startingRole) {
+            query.roleId = { gt: startingRole };
+        }
+
+        const totalCount = await this._client.roleAssignment.count({
+            where: simpleQuery,
+        });
+        const assignments = await this._client.roleAssignment.findMany({
+            where: query,
+            select: {
+                roleId: true,
+                type: true,
+                subjectId: true,
+                expireTime: true,
+            },
+            take: 10,
+            orderBy: [
+                {
+                    recordName: 'asc',
+                },
+                {
+                    roleId: 'asc',
+                },
+                {
+                    subjectId: 'asc',
+                },
+            ],
+        });
+
+        const results = assignments.map((a) => {
+            if (a.type === 'inst') {
+                return {
+                    type: 'inst',
+                    inst: a.subjectId,
+                    role: {
+                        role: a.roleId,
+                        expireTimeMs: getExpireTime(
+                            convertToMillis(a.expireTime)
+                        ),
+                    },
+                } as const;
+            } else {
+                return {
+                    type: 'user',
+                    userId: a.subjectId,
+                    role: {
+                        role: a.roleId,
+                        expireTimeMs: getExpireTime(
+                            convertToMillis(a.expireTime)
+                        ),
+                    },
+                } as const;
+            }
+        });
+
+        return {
+            assignments: results,
+            totalCount: totalCount,
         };
     }