+++ src/aux-server/aux-web/aux-auth/site/AuthRecordsFiles/AuthRecordsFiles.ts
@@ -1,0 +1,105 @@
+import { EventBus } from '@casual-simulation/aux-components';
+import Vue from 'vue';
+import Component from 'vue-class-component';
+import { Prop, Provide, Watch } from 'vue-property-decorator';
+import { authManager } from '../../shared/index';
+import { SvgIcon } from '@casual-simulation/aux-components';
+import DataSize from '../DataSize/DataSize';
+import AuthMarker from '../AuthMarker/AuthMarker';
+import { LoadingHelper } from '../LoadingHelper';
+import type {
+    ListFilesSuccess,
+    ListedFile,
+} from '@casual-simulation/aux-records';
+
+const PAGE_SIZE = 10;
+
+@Component({
+    components: {
+        'svg-icon': SvgIcon,
+        'auth-marker': AuthMarker,
+        'data-size': DataSize,
+    },
+})
+export default class AuthRecordsFiles extends Vue {
+    private _helper: LoadingHelper<ListFilesSuccess['files'][0]>;
+
+    @Prop({ required: true })
+    recordName: string;
+
+    loading: boolean = false;
+    items: {
+        mdCount: number;
+        mdPage: number;
+        startIndex: number;
+        endIndex: number;
+        mdData: ListFilesSuccess['files'];
+    } = {
+        mdCount: 0,
+        mdPage: 0,
+        mdData: [],
+        startIndex: 0,
+        endIndex: 0,
+    };
+
+    @Watch('recordName', {})
+    onRecordNameChanged(last: string, next: string) {
+        if (last !== next) {
+            this._reset();
+        }
+    }
+
+    mounted() {
+        this._reset();
+    }
+
+    private _reset() {
+        this._helper = new LoadingHelper(async (lastItem) => {
+            let result = await authManager.listFiles(
+                this.recordName,
+                lastItem?.fileName
+            );
+
+            if (result) {
+                return {
+                    items: result.files,
+                    totalCount: result.totalCount,
+                };
+            } else {
+                return {
+                    items: [],
+                    totalCount: 0,
+                };
+            }
+        });
+        this.items = {
+            mdCount: 0,
+            mdPage: 0,
+            mdData: [],
+            startIndex: 0,
+            endIndex: 0,
+        };
+        this.loading = false;
+        this.updatePagination(1, PAGE_SIZE);
+    }
+
+    changePage(change: number) {
+        this.updatePagination(this.items.mdPage + change, PAGE_SIZE);
+    }
+
+    async updatePagination(page: number, pageSize: number) {
+        let nextPage = await this._helper.loadPage(page, pageSize);
+        if (nextPage) {
+            this.items = nextPage;
+        }
+        return true;
+    }
+
+    async deleteFile(item: ListedFile) {
+        if (await authManager.eraseFile(this.recordName, item.url)) {
+            this.items.mdData = this.items.mdData.filter(
+                (i) => i.url !== item.url
+            );
+        }
+    }
+}